# doc/api.yaml
openapi: 3.0.0
info: 
  description: WASAPhoto REST APIs
  title: WASAPhoto APIs
  version: 1.0.0
servers:
  - url: http://localhost:3000
...
security:
  - BasicAuth: []
  - BearerAuth: []
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    banninguserid:
      schema:
        description: unique id related to user who bans
        type: integer
        example: 1
        readOnly: true
      name: banninguserid
      in: path
      required: true
      description: User ID who bans
    banneduserid:
      schema:
        description: unique id related to user who banned
        type: integer
        example: 1
        readOnly: true
      name: banneduserid
      in: path
      required: true
      description: User ID who bans
...
    userid:
      schema:
        description: unique id related to user
        type: integer
        example: 1
        readOnly: true
      name: userid
      in: path
      required: true
      description: User ID
...
  responses:
    BadRequest:
      description: |-
        The request was not compliant with the documentation
        (eg. missing fields, etc).
    ForbiddenRequest:
      description: |-
        Clien executes not request to the contents without
        required permissions.    
    InternalServerError:
      description: |-
        The server encountered an internal error.
        Further info in server logs.
    UnauthorizedRequest:
      description: |-
        The request had been executed without login
...
  schemas:
...
    DataAccountUpdated:
      title: user's account
      description: represents user's account modifications
      type: object
      properties:
        entity:
          description: entity which is represented by user
          enum: ["username","password","avatar"]
          example: "username"
        
         
      example:
        {
          entity: username
          
        }
    DataAccountUpdate:
      title: account update
      description: represents request body for account update
      type: object
      properties:
        newValue:
          description: new value for a component user wants change
          type: string
          example: newusername
          minLength: 3
          maxLength: 1000000
          pattern: '^.*?$'
          #can be a binary string and then u can check data inside
      example:
        {
          newValue: somenickname
        }
...
    banToUser:
      title: ban user
      description: ban user entity
      type: object
      properties:
        banningUser: 
          description: Id of user who bans
          type: integer
          example: 2
        bannedUser:
          description: Id of banned user
          type: integer
          example: 7
      example:
        {
          banningUser: 2,
          bannedUser: 1
        }
        
    #there must be a request bodies for all components
...
paths:
...
  /profiles/{banninguserid}/banuser/{banneduserid}:
  
    parameters:
      - $ref: "#/components/parameters/banninguserid"
      - $ref: "#/components/parameters/banneduserid"
    put:
      operationId: banUser
      tags:
      - Account
      summary: ban User without giving possibility to see posts
      description: |-
        addition banned user to collection of banned users by a certain one
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/banToUser" }
      responses:
        "204":
           description: Ban entity had been added successfully
          
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedRequest" }
        "403": { $ref: "#/components/responses/ForbiddenRequest" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    delete:
      operationId: unbanUser
      tags:
        - Account
      security:
        - BearerAuth: []
      summary: unban User and give possibility to see posts
      description: |-
        Removes a pair (userBans, userBanned) from a database
        after that operation banned user can see posts of userBans
      responses:
        "204":
          description: Ban entity was deleted successfully
#        "200":
#          description: Entity for banning had been deleted
#          content:
#            application/json:
#              schema: { $ref: "#/components/schemas/userBanChanged" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedRequest" }
        "403": { $ref: "#/components/responses/ForbiddenRequest" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
  /profiles/{userid}/changePassword:
    parameters:
      - $ref: "#/components/parameters/userid"
    patch:
      operationId: changePassword
      tags:
      - Account
      summary: Modifies password of user
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DataAccountUpdate" }
         
      description: |-
        Changes user's password
      responses:
        "200":
          description: user password had been changed successfully
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DataAccountUpdated" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedRequest" }
        "403": { $ref: "#/components/responses/ForbiddenRequest" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /profiles/{userid}/changeAvatar:
    parameters:
      - $ref: "#/components/parameters/userid"
    patch:
      operationId: changeAvatar
      tags:
      - Account
      summary: Changes avatar of user
      description: |-
        Changes user's avatar
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DataAccountUpdate" }
      responses:
        "200":
          description: user's avatar had been changed successfully
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DataAccountUpdated" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedRequest" }
        "403": { $ref: "#/components/responses/ForbiddenRequest" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /profiles/{userid}/deleteAccount:
    parameters:
      - $ref: "#/components/parameters/userid"

    delete:
      operationId: deleteAccount
      tags:
      - Account
      summary: Delete Account of user
      description: |-
        Delete account from application
      responses:
        204:
          description:  user's profile had been deleted successfully.
#        "200":
#          description: user's profile had been deleted successfully
#          content:
#            application/json:
#              schema: { $ref: "#/components/schemas/DataAccountUpdated" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedRequest" }
        "403": { $ref: "#/components/responses/ForbiddenRequest" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...